Created UI-Loading-Model-Chatting-LLM
Per title of this stage commit, a basic ui has been created, we are able to manually load the model, and chat with the ai. Minor ui clean is needed for basic ui is feature complete. Also needs a settings, and about page. Function calling not implemented yet. WebWorker is in place to replace service worker.